Move customer import normalization out of route
